+++ Sources/KeyboardKit/Views/System/SystemKeyboardActionButtonContent.swift
@@ -12,6 +12,9 @@ import SwiftUI
  This view resolves the correct content for a certain action,
  which will result in either a text or image, or nothing. It
  also sets up line limits, vertical offsets etc.
+ 
+ This view will also setup a special spacebar that shows the
+ locale name if the `context` has more than a one locale.
  */
 public struct SystemKeyboardActionButtonContent: View {
     
@@ -21,26 +24,30 @@ public struct SystemKeyboardActionButtonContent: View {
      - Parameters:
        - action: The action for which to generate content.
        - appearance: The appearance to apply to the content.
+       - context: The context to use when resolving content.
      */
     public init(
         action: KeyboardAction,
-        appearance: KeyboardAppearance) {
+        appearance: KeyboardAppearance,
+        context: KeyboardContext) {
         self.action = action
         self.appearance = appearance
+        self.context = context
     }
     
     private let action: KeyboardAction
     private let appearance: KeyboardAppearance
+    private let context: KeyboardContext
     
     public var body: some View {
         if action == .nextKeyboard {
             NextKeyboardButton()
+        } else if action == .space {
+            spaceView
         } else if let image = buttonImage {
             image
         } else if let text = buttonText {
-            SystemKeyboardButtonText(
-                text: text,
-                action: action)
+            textView(for: text)
         } else {
             Text("")
         }
@@ -56,19 +63,64 @@ private extension SystemKeyboardActionButtonContent {
     var buttonText: String? {
         appearance.buttonText(for: action)
     }
+    
+    var spaceView: some View {
+        SystemKeyboardSpaceButtonContent(
+            localeText: localeText,
+            spaceText: spaceText
+        )
+    }
+    
+    func textView(for text: String) -> some View {
+        SystemKeyboardButtonText(
+            text: text,
+            action: action)
+    }
+}
+
+private extension SystemKeyboardActionButtonContent {
+    
+    var localeName: String {
+        context.locale.localizedLanguageName ?? ""
+    }
+    
+    var localeText: String {
+        shouldShowLocaleName ? localeName : spaceText
+    }
+    
+    var shouldShowLocaleName: Bool {
+        context.locales.count > 1
+    }
+    
+    var spaceText: String {
+        action.standardButtonText(for: context) ?? ""
+    }
 }
 
 struct SystemKeyboardButtonContent_Previews: PreviewProvider {
     
-    static func preview(for action: KeyboardAction) -> some View {
+    static let multiLocaleContext: KeyboardContext = {
+        var context = KeyboardContext(controller: .shared)
+        context.locales = [
+            KeyboardLocale.english.locale,
+            KeyboardLocale.swedish.locale]
+        return context
+    }()
+    
+    static func preview(
+        for action: KeyboardAction,
+        multiLocale: Bool = false) -> some View {
         SystemKeyboardActionButtonContent(
             action: action,
-            appearance: .preview)
+            appearance: .preview,
+            context: multiLocale ? multiLocaleContext : .preview)
     }
     
     static var previews: some View {
         HStack {
             preview(for: .backspace)
+            preview(for: .space, multiLocale: false)
+            preview(for: .space, multiLocale: true)
             preview(for: .character("PascalCased"))
             preview(for: .character("lowercased"))
         }